backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
pydantic
python-dotenv
PyPDF2
python-docx
httpx